Fix badly written tests
